requirements.txt
dnspython==2.0.0
kafka-python==2.0.2
pymongo==3.11.0
